Create README.md
In August 2022, I began the development of this site in CodeSandbox for my Web Design I course, and at this time, still work primarily in that editor, though I've now ported the project into this repository with GitHub in order to begin versioning, not distract from lesson examples, and prevent breaking code there.  The site itself, https://thesubtlemadness.com, will go live in December as the semester comes to a close.  That will also mark the launch of our pre-sale  campaign (which also debuts the lead single) of our journey of an album, Zero Sum.  This record is the culmination of five years' work, dedication, and, at times, tumultuous ride...